feat: homeId 영구저장 및 존재하는 room 불러오는 api 완성
SharedPreferencesUtil 에 homeId 저장
RoutineViewModel의 fetchRooms을 사용하여 현재 평면도에 존재하는 방리스트를 가져옴
CreateRoutineSecondScreen에 추가

Resolves: S13P21D208-242

+++ frontend/app/src/main/java/com/example/eeum/data/model/response/floorplans/FloorPlansList.kt
@@ -3,8 +3,8 @@ package com.example.eeum.data.model.response.floorplans
 data class FloorPlansList(
     val createdAt: String,
     val floorplanId: Int,
-    val floorplansX: Int,
-    val floorplansY: Int,
+    val floorplansX: Double,
+    val floorplansY: Double,
     val homeId: Int,
     val imageUrl: String,
     val square: Double,
